make default constructed comm invalid
The default constructor used to duplicate the MPI_COMM_WORLD, which is a collective operation (although that is not explicitly documented).
Thus it was not possible to clone an object that uses a different communicator than world, as that requires creating an intermediate default object.

Now default constructed communicator should only be used to copy/move a valid communicator into it. For convenience a static member to create a communicator using MPI_COMM_WORLD has been added.